Added ECB payload & block size calculation utils
* Refactor BLOCK_SIZE to be usize instead of u8 (no more .into() everywhere)
* util functions added
  * determine_payload_length
  * determine_prefix_block_count
  * determine_oracle_block_size (renamed)

+++ src/lib.rs
@@ -10,7 +10,7 @@ pub mod util;
 pub mod pkcs;
 pub mod oracle;
 
-pub const BLOCK_SIZE: u8 = 16;
+pub const BLOCK_SIZE: usize = 16;
 pub type AesKey = [u8; 16];
 
 use rand::prelude::*;
